requirements.txt
Flask~=2.2.2
flask_sqlalchemy
flask_login
flask_wtf
WTForms~=3.0.1
email_validator
python-dotenv~=0.21.0
bcrypt~=4.0.1
cryptography==38.0.3
SQLAlchemy==1.4.44
pyotp~=2.7.0
MarkupSafe~=2.1.1